+++ Sources/Navigator/EPUB/Preferences/EPUBPreferences.swift
@@ -107,9 +107,6 @@ public struct EPUBPreferences: ConfigurablePreferences, Sendable {
     /// Reader theme.
     public var theme: Theme?
 
-    /// Scale applied to all element font sizes.
-    public var typeScale: Double?
-
     /// Indicates whether the text should be laid out vertically.
     ///
     /// This is used for example with CJK languages. This setting is
@@ -148,7 +145,6 @@ public struct EPUBPreferences: ConfigurablePreferences, Sendable {
         textColor: Color? = nil,
         textNormalization: Bool? = nil,
         theme: Theme? = nil,
-        typeScale: Double? = nil,
         verticalText: Bool? = nil,
         wordSpacing: Double? = nil
     ) {
@@ -180,7 +176,6 @@ public struct EPUBPreferences: ConfigurablePreferences, Sendable {
         self.textColor = textColor
         self.textNormalization = textNormalization
         self.theme = theme
-        self.typeScale = typeScale.map { max($0, 0) }
         self.verticalText = verticalText
         self.wordSpacing = wordSpacing.map { max($0, 0) }
     }
@@ -215,7 +210,6 @@ public struct EPUBPreferences: ConfigurablePreferences, Sendable {
             textColor: other.textColor ?? textColor,
             textNormalization: other.textNormalization ?? textNormalization,
             theme: other.theme ?? theme,
-            typeScale: other.typeScale ?? typeScale,
             verticalText: other.verticalText ?? verticalText,
             wordSpacing: other.wordSpacing ?? wordSpacing
         )
@@ -223,6 +217,11 @@ public struct EPUBPreferences: ConfigurablePreferences, Sendable {
 
     @available(*, unavailable, message: "Not needed anymore with Readium CSS v2, user settings are applied as soon as they are set")
     public var publisherStyles: Bool? {
+        fatalError()
+    }
+
+    @available(*, unavailable, message: "Not available in Readium CSS v2")
+    public var typeScale: Double? {
         fatalError()
     }